Fill screen black upon keypress
Still to implement: unset the black screen if the user lets go of the keyboard.

The way I will do this is the opposite of the draw loop - I will set the screen registers to white AFTER the draw loop, in any case.  The only difference is that in each iteration of the scan loop - which will set the registers back to white - it will have to break and start the draw loop if there is a keypress.