What plan? DONE
Had an idea while doing my other program/critiquing it.

Basically, it was getting too hairy. I was checking here, there, and everywhere what the current state was etc.

In this new way, I just have a main loop that tracks the position of the screen, and goes through every register whether the other bits of the program like it or not. Then, in each iteration, we either set the current register black or white depending on the keypress.

Instead of going back to A each time we detect a change, and detecting changes constantly, we just CONTINUE doing what we're doing, and slightly change the execution if there is/isn't input.

Way more responsive. Done.